+++ composer.json
@@ -11,9 +11,12 @@
 	"require": {
 		"php": ">=5.4.0"
 	},
+	"require-dev": {
+		"phpunit/phpunit": "^5"
+	},
 	"autoload": {
 		"psr-4": {
-			"FormulaParser\\": ""
+			"FormulaParser\\": "src/"
 		}
 	}
 }
